+++ lang/en/block_people.php
@@ -44,6 +44,9 @@ $string['setting_resetvisibility'] = 'Reset visibility';
 $string['setting_resetvisibility_desc'] = 'By enabling this checkbox, the visibility of all existing "block_people" instances will be set to visible (again).<br/>
 Please note: <br/>
 After saving this option, the database operations for resetting the visibility will be triggered and this checkbox will be unticked again. The next enabling and saving of this feature will trigger the database operations for resetting the visibility again. ';
+$string['setting_rolesheading'] = 'Roles';
+$string['setting_roles'] = 'Show these roles';
+$string['setting_roles_desc'] = 'This setting allows you to control which users appear in this block. Users need to have at least one of these roles in a course to be shown in the block.';
 
 // Notifications.
 $string['resetvisibilitysuccess'] = 'Success! All "block_people" instances are visible (again). <br/> The setting "Reset visibility" has been reset.';