+++ settings.php
@@ -29,6 +29,19 @@ if ($ADMIN->fulltree) {
     // Locallib for updatedcallback function.
     require_once($CFG->dirroot.'/blocks/people/locallib.php');
 
+    // Settings title to group role related settings together with a common heading. We don't want a description here.
+    $name = 'block_people/rolesheading';
+    $title = get_string('setting_rolesheading', 'block_people', null, true);
+    $setting = new admin_setting_heading($name, $title, null);
+    $settings->add($setting);
+
+    // Setting to configure the roles to be shown within the block.
+    $name = 'block_people/roles';
+    $title = get_string('setting_roles', 'block_people', null, true);
+    $description = get_string('setting_roles_desc', 'block_people', null, true);
+    $default = array('editingteacher');
+    $settings->add(new admin_setting_pickroles($name, $title, $description, $default));
+
     // Settings title to group partictpants page related settings together with a common heading. We don't want a description here.
     $name = 'block_people/participantspageheading';
     $title = get_string('setting_participantspageheading', 'block_people', null, true);